cdk/module/aurora_mysql/requirements.txt
aws-cdk-lib==2.16.0
constructs>=10.0.0,<11.0.0
PyYAML==5.4.1
mockito==1.3.0
Jinja2==3.0.3
jsonschema
boto3>=1.21.45